Uninitialized string in heartbeat/findif.c
Patch from Pavol Gono to make findif a little more robust in
the case that there are no interfaces present.

http://permalink.gmane.org/gmane.linux.highavailability.devel/1978

I guess that the snprintf call's return value should be checked
to see if the buffer was big enough. However, this is a little
tricky as the return value of snprintf depends on the libc
(and weather or not it's snprintf is C99 compliant or not).
Given that the buffer unlikely to be too short, this seems to
be a minor concern.

Cc: Pavol Gono <palo.gono@gmail.com>
Signed-off-by: Simon Horman <horms@verge.net.au>