Novell bug 250207: SLES10 SP1 beta4 x86_64 problems with Heartbeat 2.0.8
The low level heartbeat connectivity code exchanges lists of configured
nodes when a node joins.  This list of nodes should not have contained
any ping nodes, but it did.  If the list of ping nodes disagreed between
two machines then very weird things could happen involving having
"missing" ping nodes being automaticallly added as regular nodes on
machines which didn't have them configured.  And then all kinds of
strangeness happened after that which meant that the only way to fix
this was to make things consistent INCLUDING the hostcache files.

Note that this would also cause problems on machines where the
ping nodes were configured differently deliberately.  In the past,
the only thing that really mattered was the _number_ of the ping
nodes - not their names/addresses.

This problem should also go away with this patch.