Bug 250207 - SLES10 SP1 beta4 x86_64 problems with Heartbeat 2.0.8 This change does two things:  1) It makes heartbeat retry starting the CCM in the case that the CCM     can't reconcile the configured set of nodes in the cluster with     some other node.
 2) It prints out a little better explanation of where to start looking
    to diagnose this problem.

As far as (1) is concerned:
If the problem is some kind of transient timing weirdness, this will make it
recover.

Heartbeat will retry running it something like ten times before it gives up.
That's what changing the exit value from 10 to 1 (or any value != 10) does.
The exit value and how many times heartbeat retries are unrelated.