LRM (bug 1594): change handling of cancel, resource delete, and flush ops requests.
These requests, though asynchronous in nature because of the
possibility that a process could be running at the time of the
request, now appear synchronous to the client. In case there was
no process running for the operation or the resource, reply is
sent immediately back. Otherwise, we send the response to the
client as soon as the process has exited. This has been described
in include/lrm/lrm_api.h.

Handling of operations has been thoroughly examined and some
glitches found and fixed. Some previous patches (see
10631:5f40fa430610) are obsoleted by this patch. Operations
handling flow has been simplified (see calls to on_op_done).

Operations are not allowed to hang around without its resource.
Logging has been updated accordingly and a resource missing is
considered an error or critical condition.