Remove autoconf and friends on make distclean I'm not sure what the intention of the symlinks to autoconf, autoheader and automake added by 2d04e4dd3a90857e280e3036f297b91ab0355138 are, but it seems reasonable to remove them on make distclean as they aren't present in a dist tarball.
This actually fixes a problem when building debian packages.
Basically if ./ConfigureMe is run for any reason when
updating a debian package, say to effect a change to a Makefile.am,
then these symlinks are created, and the debian packaging
system can't cope with this.

If this change is not acceptable I can just push it into debian/rules,
but I think that generally its nicer not to push things into
packaging specific portions of the build system.

I'm am looking this patch being included post 2.1.1.1

Signed-off-by: Simon Horman <horms@verge.net.au>