configure.in: fix for platforms with libc in /lib/tls
The changeset 5f8f217f306a introduced looking for a "proper
library suffix" probably in order to deal better with 64-bit
platforms (it's the one in /usr/lib/heartbeat for example).
However, it fails in case libc is linked from /lib/tls/libc.so,
producing "tls" as a library suffix, which is obviously wrong.