[lvs-users] Patch for ldirectord when using mysql service My first go at this mailing list, hope this is the place to post a patch...
We're using a 2-server-cluster LVS and HA setup to have a loadbalanced
high available MySQL database server.
Currently running CentOS 5.0, MySQL 5.0.22 and the heartbeat packages
version 2.1.2.

We've been having some trouble getting ldirectord doing it's work when
it comes to adding real servers to the VIP host. The real-servers failed
due to errors in the ldirectord when checking the database with a
provided SELECT query. To get this working, we had to fix 2 very small
problem in ldirectord, in the check_sql method.

Ldirectord version string is "Linux Director v1.186-ha-2.1.2"

Cc: Thomas Boehme <TBoehme@studivz.net>
Cc: Sverre Marvik <sm@mobiletech.no>
Signed-off-by: Simon Horman <horms@verge.net.au>