stonithd: shuffle dropping privileges around again
All message handlers (for both heartbeat and IPC messages) are
run as root. The code looks as if it should've been the other way
around, i.e. every call to a stonith agent is wrapped in
return_to_orig_privs/return_to_dropped_privs. This patch reverts
the previous one and raises privileges around the code where the
hostlist is handled. This should be further investigated.