High: Tools: attrd - Prevent race condition resulting in the cluster forgetting node's wish to shut down
Multiple updates can queue up if attrd takes a long time to connect to the CIB.
Due to the way mainloop is used, network updates are applied before local ones.
If the node requests shutdown before the CIB connection is established, then the
request will be overwritten by the local node resetting any existing values
(both requests being queued).

The solution is to activate the mainloop before connecting to the CIB and
retry connections at a lower priority than updates.