High: ais: Bug rhbz#525552 - Move non-threadsafe calls to setenv() to after the fork()
Only the child processes actually needs these values and after the
call to fork we're no longer multi-threaded, so it is safe to set
them.

This prevents segfaults caused when one thread calls getenv(), via
syslog() and tzset(), at the same time that another is calling
setenv().  See https://bugzilla.redhat.com/show_bug.cgi?id=561282 for
more details.

The glibc maintainers refuse to acknowledge the issue as they're
apparently on a crusade to rid the world of threads, so instead we're
stuck with work-arounds like this.